requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-WTF>=1.2.1
WTForms>=3.0.1
Werkzeug>=3.0.0
click>=8.1.7
itsdangerous>=2.1.2
jinja2>=3.1.4
openpyxl>=3.1.5
python-docx>=1.1.0
reportlab>=4.2.5

